requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements/base.in
#
anyio==4.11.0
    # via httpx
asgiref==3.10.0
    # via
    #   django
    #   django-cors-headers
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
cryptography==46.0.3
    # via -r requirements/base.in
django==5.2.7
    # via
    #   -r requirements/base.in
    #   django-cors-headers
    #   djangorestframework
    #   drf-spectacular
django-cors-headers==4.4.0
    # via -r requirements/base.in
djangorestframework==3.16.1
    # via
    #   -r requirements/base.in
    #   drf-spectacular
drf-spectacular==0.27.2
    # via -r requirements/base.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.25.2
    # via -r requirements/base.in
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.25.1
    # via drf-spectacular
jsonschema-specifications==2025.9.1
    # via jsonschema
marshmallow==3.26.1
    # via -r requirements/base.in
packaging==25.0
    # via marshmallow
pycparser==2.23
    # via cffi
python-decouple==3.8
    # via -r requirements/base.in
pyyaml==6.0.3
    # via drf-spectacular
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via -r requirements/base.in
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlparse==0.5.3
    # via django
typing-extensions==4.15.0
    # via
    #   anyio
    #   referencing
uritemplate==4.2.0
    # via drf-spectacular
urllib3==2.5.0
    # via requests
